Menubar initalization changed.
Moved most of the init code for the menu into an own class to be able to
remove the PMShell pointer from init in PMPart. There it is only needed
to access the menu. So why include all classes needed by PMShell.
Its a matter of encapsulation to make it more clear here.